+++ lib/parse/stack/generators/templates/model_installation.rb
@@ -1,0 +1,6 @@
+
+
+class Parse::Installation < Parse::Object
+  # See: https://github.com/modernistik/parse-stack#parseinstallation
+  # add additional properties here
+end