+++ lib/parse/stack/generators/templates/model_session.rb
@@ -1,0 +1,6 @@
+
+
+class Parse::Role < Parse::Object
+  # See: https://github.com/modernistik/parse-stack#parsesession
+  # add additional properties here
+end