+++ lib/parse/stack/generators/templates/parse.rb
@@ -1,0 +1,13 @@
+require 'parse/stack'
+
+# Set your specific Parse keys in your ENV. For all connection options, see
+# https://github.com/modernistik/parse-stack#connection-setup
+
+Parse.setup app_id: ENV['PARSE_APP_ID'],
+           api_key: ENV['PARSE_API_KEY'],
+        server_url: 'https://api.parse.com/1/',
+          # optional
+        master_key: ENV['PARSE_MASTER_KEY'],
+           logging: false,
+             cache: Moneta.new(:File, dir: 'tmp/cache'),
+           expires: 5 # cache ttl 3 seconds