+++ Changes.md
@@ -2,9 +2,11 @@
 
 1.4.0
 -----------
+- NEW: Support for rails generators: `parse_stack:install` and `parse_stack:model`.
 - Support for calling `before_save` and `before_destroy` callbacks in your model when a Parse::Object is returned by your `before_save` or `before_delete` webhook respectively.
 - Parse::Query `:cache` expression now allows integer values to define the specific cache duration for this specific query request. If `false` is passed, will ignore the cache and make the request regardless if a cache response is available. If `true` is passed (default), it will use the value configured when setting up when calling `Parse.setup`.
 - Fixes the use of `:use_master_key` in Parse::Query.
+- Fixes to the cache key used in middleware.
 
 1.3.8
 -----------